Moved all common api classes to api.commons, and used one shared session var
