+++ splitio/api/commons.py
@@ -1,9 +1,11 @@
 """Commons module."""
+from collections import namedtuple
 from splitio.util.time import get_current_epoch_time_ms
 
 _CACHE_CONTROL = 'Cache-Control'
 _CACHE_CONTROL_NO_CACHE = 'no-cache'
 
+HttpResponse = namedtuple('HttpResponse', ['status_code', 'body'])
 
 def headers_from_metadata(sdk_metadata, client_key=None):
     """
@@ -114,3 +116,40 @@ def build_fetch(change_number, fetch_options, metadata):
     if fetch_options.change_number is not None:
         query['till'] = fetch_options.change_number
     return query, extra_headers
+
+class APIException(Exception):
+    """Exception to raise when an API call fails."""
+
+    def __init__(self, custom_message, status_code=None):
+        """Constructor."""
+        Exception.__init__(self, custom_message)
+        self._status_code = status_code if status_code else -1
+
+    @property
+    def status_code(self):
+        """Return HTTP status code."""
+        return self._status_code
+
+class HttpClientException(Exception):
+    """HTTP Client exception."""
+
+    def __init__(self, message):
+        """
+        Class constructor.
+
+        :param message: Information on why this exception happened.
+        :type message: str
+        """
+        Exception.__init__(self, message)
+
+def build_basic_headers(apikey):
+    """
+    Build basic headers with auth.
+
+    :param apikey: API token used to identify backend calls.
+    :type apikey: str
+    """
+    return {
+        'Content-Type': 'application/json',
+        'Authorization': "Bearer %s" % apikey
+    }
