+++ splitio/api/telemetry.py
@@ -1,9 +1,8 @@
 """Impressions API module."""
 import logging
 
-from splitio.api import APIException
 from splitio.api.client import HttpClientException
-from splitio.api.commons import headers_from_metadata, record_telemetry
+from splitio.api.commons import headers_from_metadata, record_telemetry, APIException
 from splitio.util.time import get_current_epoch_time_ms
 from splitio.models.telemetry import HTTPExceptionsAndLatencies
